fix: type casting

+++ lightwood/analysis/nc/nc.py
@@ -128,7 +128,7 @@ class MarginErrFunc(ClassificationErrFunc):
         super(MarginErrFunc, self).__init__()
 
     def apply(self, prediction, y):
-        prediction = deepcopy(prediction)
+        prediction = deepcopy(prediction).astype(float)
         prob = np.zeros(y.size, dtype=np.float32)
         for i, y_ in enumerate(y):
             if y_ >= prediction.shape[1]:
